+++ src/app/components/index.ts
@@ -1,0 +1,5 @@
+export * from "./banner"
+export * from "./layout"
+export * from "./popular-categories"
+export * from "./theme-provider"
+export * from "./top"
