+++ src/app/components/top/index.tsx
@@ -4,7 +4,7 @@ import { Author, Book } from "@/api/types"
 import { BooksTop } from "@/app/components/top/books-top"
 import { AuthorsTop } from "@/app/components/top/authors-top"
 
-const Top: FC<{ books: Book[]; authors: Author[] }> = ({ books, authors }) => {
+export const Top: FC<{ books: Book[]; authors: Author[] }> = ({ books, authors }) => {
   return (
     <Grid container pb={3}>
       <Grid item xs={12} sm={6} md={8}>
@@ -16,5 +16,3 @@ const Top: FC<{ books: Book[]; authors: Author[] }> = ({ books, authors }) => {
     </Grid>
   )
 }
-
-export default Top
